requirements.txt
# discord dependencies
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
disnake==2.9.0
frozenlist==1.4.0
idna==3.4
install==1.3.5
multidict==6.0.4
yarl==1.9.2

# env controlls
python-decouple